Add workflow to publish one lib per tag to gh-pages branch (#11)
The workflow runs whenever a tag gets added to the repository. It updates the
gh-pages branch as needed with new versions of the JS library:
- one file per version, e.g. `i-slide-1.0.0.js`
- one file per major version, e.g. `i-slide-1.js`
- the nightly version, named `i-slide-nightly.js`
- the latest released (i.e. tagged) version, named `i-slide.js`

Workflow runs whenever a push gets made to the main branch and whenever a tag gets added.

It only takes semver tags into account.

Workflow also updates the README and LICENSE documents in the gh-pages branch if needed.