Prevent overflows and flag as block element for PDF as well
This adds back the `overflow: hidden` directive, coupled with a
`position: relative` directive to make sure that absolutely positioned elements
respect the overflow.

The `:host` styles were not set in the PDF case, which meant that the custom
element was rendered as an inline element. This did not have any consequence
whatsoever, but the custom element should be a block element regardless of the
type of slides that get displayed.

From a testing perspective, I'm not really sure how to access these properties.
There are set on the `:host`, which is neither the element itself, nor the
shadow root, so I don't see any immediate way to `querySelector` the right
"element" to read the values I'm interested in...