Avoid overflows, rewrite test framework and add related tests
To avoid overflows, `position: relative; overflow: hidden` directives get set on
the root element of the shadow root (so that they cannot be affected by CSS
properties set on the <i-slide> element).

This update rewrites the test framework to:
- make it possible to test that <i-slide> is displayed as a block element
- make it possible to test that the right directives are set on elements in
the shadow root, and more generally speaking execute arbitrary evaluation
functions.
- generalize the path to a proper `querySelector` string (with a couple of
extensions to select the attribute)
- slightly simplify the structure of the `tests` array.

Tests completed with a few tests on the display, position and overflow
directives.