Prevent overflows and flag as block element for PDF as well (#21)
The `:host` styles were not set in the PDF case, which meant that the custom
element was rendered as an inline element. This did not have any consequence
whatsoever, but the custom element should be a block element regardless of the
type of slides that get displayed.

To avoid overflows, `position: relative; overflow: hidden` directives get set on
the root element of the shadow root (so that they cannot be affected by CSS
properties set on the <i-slide> element).

This update rewrites the test framework to:
- make it possible to test that <i-slide> is displayed as a block element
- make it possible to test that the right directives are set on elements in
the shadow root, and more generally speaking execute arbitrary evaluation
functions.
- generalize the path to a proper `querySelector` string (with a couple of
extensions to select the attribute)
- slightly simplify the structure of the `tests` array.

Tests completed with a few tests on the display, position and overflow
directives.