Set :host width as well
As a block element, the `<i-slide>` was taking the entire width available,
even though its content width is clamped to the value of the `width` attribute
(or 300px). With this update, the width of the element is explicitly set to the
width of its content. That is more consistent with the way `<img>` and
`<iframe>` work.