Set aria-busy attribute while loading slides
The attribute gets set when slides are being fetched, and reset when the slide
is rendered.

This update also fixes a minor glitch in the test framework that made it add
an extraneaous "}" character in the HTML of the test page.

More importantly, this update also fixes a re-entrancy issue within the
attribute changed callback, using a "busyCounter" to know when "render" can
be called (and when "aria-busy" can be reset). Tests don't cover that part,
coz' that's harder to test...