Only fire the "load" event when no more busy work
If the "src" attribute gets changed multiple times in a row, this prevents the
"load" event from being fired once for each fetch-and-render phase. The "load"
event will only be fired when the final fetch-and-render phase will have been
completed.

(The logic may be improved. For once, we probably would prefer not to run the
render phase if another fetch is being performed that will replace the content)