Handle fetch/render cycles, property/attr reflection, aria-busy, and fire "load" only once (#28)
This rewrites most of the logic in preparation to reacting to box resizing
events. Main changes:
- The ISlide class now uses private data and methods not to expose inner logic
to the external world
- JS attributes that reflect HTML attributes now use getters/setters, and return
values that "make sense" (`src` returns an absolute URL, `width` an integer)
- Relative URLs get resolved against the right base URL
- Updates to JS properties get propagated to the relevant HTML attributes
- Updates to properties/attributes that require a new fetch-and-render cycle
schedule that fetch-and-render cycle for next tick (with at most one cycle
scheduled per tick to account for situations where multiple attributes get
changed during the same tick)
- Fetch-and-render cycles get canceled as soon as possible when they should not
be finished because another cycle needs to run
- "aria-busy" and "load" are reset and fired at the end of a cycle only when
there are no other cycles.
- Render operations are serialized to make sure that the latest value of "width"
gets eventually used if that attribute changes while a rendering operation is
already running.

Tests completed to test the `aria-busy` attribute as well as to test relection of
attributes and propagation of properties to attributes.