Add height attribute and resize content automatically
The `height` attribute completes the "width" attribute and allows authors to
fully impose the size the `<i-slide>` element in the HTML. When the aspect ratio
of the element is different from the intrinsic aspect ratio of the slide, the
slide is rendered as though it was styled with:

  object-fit: contain;
  object-position: top left;

These properties should be properly supported and default would also need to
change to `fill` and `50% 50%` as defined in CSS. That is a TODO...

The contents of the `<i-slide>` element are now also automatically scaled
up/down when additional CSS dimension constraints are imposed on the element.
This gets done by using a `ResizeObserver`.

The scaling logic was moved out of `#render()`. A number of private properties
had to be added to track last known info about the slide that is rendered, e.g.
to keep a pointer to the PDF page and more importantly to keep track of the last
applied scale and target dimensions so as not to re-scale content when that is
not needed. These properties could perhaps be consolidated in a
`this.#slideInfo` object (not done here).

The render logic was serialized to avoid re-entrancy issues. That may not be
fully needed but debugging renders and resizes is already hard enough when
things are serialized...

The code now also includes a `log` function to log a few function calls to the
console. Function is a no-op by default and enabled by setting a global `DEBUG`
variable. That variable gets set when tests are run in debug mode.

Tests on the height attribute and resizing tests still need to be created.