Add tests on height attr/prop, box dimensions, scaling and resize
This completes tests on attributes and properties with checks on `height`,
and adds a number of tests on outer and inner element/shadow root dimensions
to make sure that the component behaves correctly when CSS directives impose
its dimensions or when content otherwise needs to be re-scaled.

This includes a few tests on dynamically changing the URL of the slide (it's
easy to forget to reset something when the source of the slide changes...)

The test framework was updated to check the presence of an `ISLIDETEST`
environment variable. When set, the test framework only runs tests whose
title contains the value of the ISLIDETEST variable. This makes it easier to
run a single test while developing the test suite, as well as to run a single
test in DEBUG mode.