Add height attribute and resize content automatically (#30)
The `height` attribute completes the "width" attribute and allows authors to
fully impose the size the `<i-slide>` element in the HTML. When the aspect ratio
of the element is different from the intrinsic aspect ratio of the slide, the
slide is rendered as though it was styled with:

  object-fit: contain;
  object-position: top left;

These properties should be properly supported and default would also need to
change to `fill` and `50% 50%` as defined in CSS. That is a TODO...

The contents of the `<i-slide>` element are now also automatically scaled
up/down when additional CSS dimension constraints are imposed on the element.
This gets done by using a `ResizeObserver`.

The scaling logic was moved out of `#render()`. A number of private properties
had to be added to track last known info about the slide that is rendered, e.g.
to keep a pointer to the PDF page and more importantly to keep track of the last
applied scale and target dimensions so as not to re-scale content when that is
not needed. These properties could perhaps be consolidated in a
`this.#slideInfo` object (not done here).

The render logic may not need to remain serialized but debugging renders and
resizes is already hard enough when things are serialized...

The code now also includes a `log` function to log a few function calls to the
console. Function is a no-op by default and enabled by setting a global `DEBUG`
variable. That variable gets set when tests are run in debug mode.

This also completes tests on attributes and properties with checks on `height`,
and adds a number of tests on outer and inner element/shadow root dimensions
to make sure that the component behaves correctly when CSS directives impose
its dimensions or when content otherwise needs to be re-scaled.

This includes a few tests on dynamically changing the URL of the slide (it's
easy to forget to reset something when the source of the slide changes...)

The test framework was updated to check the presence of an `ISLIDETEST`
environment variable. When set, the test framework only runs tests whose
title contains the value of the ISLIDETEST variable. This makes it easier to
run a single test while developing the test suite, as well as to run a single
test in DEBUG mode.