Always take width/height imposed by CSS into account
This update makes the code take `clientWidth` and `clientHeight` into account
when rendering the slide for the first time, to handle the situation where the
resize event gets triggered before rendering (which happens when the ISlide
element is rendered as a block with explicit `width` and `height` CSS
properties).

A non-regression test was added to the test suite.

The width of one of the tests was adjusted to avoid off-by-1 fluctuating errors,
which seem to be triggered by some rounding issue when content gets rendered
(not a real code error in the sense that the content eventually takes the right
dimensions, although there may be something to improve to avoid re-scaling the
content).

Fix issue #42