Switch PDF.js CDN to jsdelivr.com, and bump version
This makes the code retrieve PDF.js from the jsdelivr.com CDN instead of from
unpkg.com because unpkg.com is currently experiencing recurring downtimes. Not
sure if jsdelivr.com will prove more reliable. Longer term, we may want to host
the code ourselves instead of relying on free CDN services.

A new minor version of PDF.js is also just out, so using that opportunity to
switch to it.