Bump version of PDF.js (and other dependencies) (#51)
* Bump version of PDF.js (and other dependencies)

This bumps the now obsolete version of PDF.js that was used and that cannot be
retrieved from unpkg.com anymore. The new version is an ES module, so
integration becomes slightly simpler. Newer versions introduced a few breaking
changes though, notably:

1. The slides container now need to have a `pdfViewer` class
2. Some options passed to the `PDFPageView` constructor evolved
3. Scaling within the PDF.js now leverages CSS variables and CSS computations
through the `calc()` and `round()` functions, from an initial 720*405 canvas.
When scale does not yield an exact float, resulting dimensions can be off by
one pixel as a result. Some tests that play with scaling had to be adjusted to
be more friendly with the initial 720*405 dimensions.

Other dev dependencies were bumped too. This was actually required for
Puppeteer because PDF.js uses newer JavaScript features that very old versions
of Chromium did not support.

Note the legacy build of PDF.js is used on purpose so as to work well across a
wider set of browsers, see:
https://github.com/mozilla/pdf.js/wiki/Frequently-Asked-Questions#faq-support

Fixes #50.

* Also bump version of Node used in jobs

Not sure why the test job failed but tests pass on my local computer, so
perhaps due to some Node.js hiccup with now old version 18.

* Disable AppArmor in jobs for Puppeteer execution

* Switch PDF.js CDN to jsdelivr.com, and bump version

This makes the code retrieve PDF.js from the jsdelivr.com CDN instead of from
unpkg.com because unpkg.com is currently experiencing recurring downtimes. Not
sure if jsdelivr.com will prove more reliable. Longer term, we may want to host
the code ourselves instead of relying on free CDN services.

A new minor version of PDF.js is also just out, so using that opportunity to
switch to it.