Override slide numbering styles to render the correct slide number
Whenever the ISlide component needs to render a slide, it creates a fake
slideset that only contains a copy of the slide to render. This is done on
purpose to avoid having to clone the whole slideset whenever a slide is
rendered (the main use case for the component is a page where all slides
are rendered, the slideset would have to be cloned for each of the slide).

This works well for PDF slidesets, but HTML slidesets typically use a CSS
rule of the form `.slide::after { content: counter(slide) }` to render the
slide number. That works well when all slides are available, but means that
the slide number ends up being "1" for all HTML slides rendered through the
ISlide component.

The right solution would be to clone the whole slideset each time, as proposed
in #49. But that's precisely what we're trying to avoid...

This update takes a different approach and rather inject a new CSS rule with
the actual slide number. For example, `.slide::after { content: "4" }`.

By definition, this solution is more brittle because it relies on inner
knowledge of how slide frameworks work. Shower and b6+ follow the same pattern
today, but that could change in the future. Other HTML slide frameworks may use
a different one. Also, the solution only accounts for slide numbers, while
frameworks may have additional features that depend on the presence of all
slides.