Fix embedded PDF slides in Chrome (#58)
Chrome does not properly strip newline characters in `data:` URLs, and creates
`%0A` characters in base64 strings embedded in `href` attributes as a result,
leading to decoding errors.

Underlying Chromium bug is tracked in:
https://issues.chromium.org/issues/40764064

This update forces a call to `decodeURI` to replace `%0A` characters with `\n`.
Note call will be a no-op in browsers such as Firefox that correctly strip
newline characters.

The demo page gets completed to advertize the capability and also provide a
handy way to test the feature. A proper test would be better but would require
additional work as the current test harness is not directly suitable for this
scenario.

Close #57.

Note: I took the opportunity to also replace occurrences of tab spaces with
spaces for consistency and to fix indentation in sane editors...